refactore code
